Update kim_main.py
Adding the "clock" vatiable with the "time.asctime()" function.

Adding second "print()"  in line "20". This prints the "clock" variable witch gives the local time and date back. With ".center(50)" the output is shown in the center of the row.


NOTE: In the future "time.asctime()" will be removed and the "clock" variable will be changed.