Features and Fix
Features:
closed #3 and #4
fix #5

Added the feature that the operating buttons can be pressed multiple times,
so the application wound crash anymore.

Also added the option for the user that he can choose if he wants to:
Close the application with writing "yes" and confirm it with enter.
Or
Restart the application with writing "restart" and confirm it with enter.